requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
dj-database-url==3.1.2
Django==5.2.13
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
environ==1.0
gunicorn==25.3.0
idna==3.13
joblib==1.5.3
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
psycopg2-binary==2.9.12
PyJWT==2.12.1
python-dateutil==2.9.0.post0
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2026.2
urllib3==2.6.3
whitenoise==6.12.0
